tracks variable config to present correct table of contents
